broke out dropdown from header into Dropdown component and implemented clickout with hooks instead of third party package
